requirements.txt
streamlit
python-dotenv
crewai
langchain
openai
